Changed to internal helper struct, which can be re-used by all wallet implementations (#3024)
Let's be done with it for now.
