+++ vcr/holder/presenter.go
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022 Nuts community
+ * Copyright (C) 2024 Nuts community
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -29,71 +29,32 @@ import (
 	ssi "github.com/nuts-foundation/go-did"
 	"github.com/nuts-foundation/go-did/did"
 	"github.com/nuts-foundation/go-did/vc"
-	"github.com/nuts-foundation/nuts-node/audit"
 	"github.com/nuts-foundation/nuts-node/auth/oauth"
-	"github.com/nuts-foundation/nuts-node/core"
 	"github.com/nuts-foundation/nuts-node/crypto"
-	"github.com/nuts-foundation/nuts-node/jsonld"
-	"github.com/nuts-foundation/nuts-node/storage"
 	"github.com/nuts-foundation/nuts-node/vcr/credential"
-	"github.com/nuts-foundation/nuts-node/vcr/credential/store"
-	"github.com/nuts-foundation/nuts-node/vcr/log"
 	"github.com/nuts-foundation/nuts-node/vcr/pe"
 	"github.com/nuts-foundation/nuts-node/vcr/signature"
 	"github.com/nuts-foundation/nuts-node/vcr/signature/proof"
-	"github.com/nuts-foundation/nuts-node/vcr/types"
-	"github.com/nuts-foundation/nuts-node/vcr/verifier"
 	"github.com/nuts-foundation/nuts-node/vdr/resolver"
-	"gorm.io/gorm"
-	"gorm.io/gorm/schema"
+	"github.com/piprate/json-gold/ld"
 	"strings"
 	"time"
 )
 
-// ErrNoCredentials is returned when no matching credentials are found in the wallet based on a PresentationDefinition
-var ErrNoCredentials = errors.New("no matching credentials")
-
-type wallet struct {
-	keyResolver   resolver.KeyResolver
-	keyStore      crypto.KeyStore
-	verifier      verifier.Verifier
-	jsonldManager jsonld.JSONLD
-	walletStore   walletStore
+type presenter struct {
+	documentLoader ld.DocumentLoader
+	keyStore       crypto.KeyStore
+	keyResolver    resolver.KeyResolver
 }
 
-// New creates a new Wallet.
-func New(
-	keyResolver resolver.KeyResolver, keyStore crypto.KeyStore, verifier verifier.Verifier, jsonldManager jsonld.JSONLD,
-	storageEngine storage.Engine) Wallet {
-	return &wallet{
-		keyResolver:   keyResolver,
-		keyStore:      keyStore,
-		verifier:      verifier,
-		jsonldManager: jsonldManager,
-		walletStore:   walletStore{db: storageEngine.GetSQLDatabase()},
-	}
-}
-
-// BuildParams contains the parameters that will be included in the signature of the verifiable presentation
-type BuildParams struct {
-	Audience string
-	Expires  time.Time
-	Nonce    string
-}
-
-func (h wallet) BuildSubmission(ctx context.Context, walletDID did.DID, presentationDefinition pe.PresentationDefinition, acceptedFormats map[string]map[string][]string, params BuildParams) (*vc.VerifiablePresentation, *pe.PresentationSubmission, error) {
-	// get VCs from own wallet
-	credentials, err := h.List(ctx, walletDID)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to retrieve wallet credentials: %w", err)
-	}
-
+func (p presenter) buildSubmission(ctx context.Context, holderDID did.DID, credentials []vc.VerifiableCredential, presentationDefinition pe.PresentationDefinition,
+	acceptedFormats map[string]map[string][]string, params BuildParams) (*vc.VerifiablePresentation, *pe.PresentationSubmission, error) {
 	// match against the wallet's credentials
 	// if there's a match, create a VP and call the token endpoint
 	// If the token endpoint succeeds, return the access token
 	// If no presentation definition matches, return a 412 "no matching credentials" error
 	builder := presentationDefinition.PresentationSubmissionBuilder()
-	builder.AddWallet(walletDID, credentials)
+	builder.AddWallet(holderDID, credentials)
 
 	// Find supported VP format, matching support from:
 	// - what the local Nuts node supports
@@ -119,7 +80,7 @@ func (h wallet) BuildSubmission(ctx context.Context, walletDID did.DID, presenta
 			return nil, nil, ErrNoCredentials
 		}
 		// add empty sign instruction
-		signInstructions = append(signInstructions, pe.SignInstruction{Holder: walletDID})
+		signInstructions = append(signInstructions, pe.SignInstruction{Holder: holderDID})
 		presentationSubmission = pe.PresentationSubmission{
 			Id:            uuid.NewString(),
 			DefinitionId:  presentationDefinition.Id,
@@ -128,7 +89,7 @@ func (h wallet) BuildSubmission(ctx context.Context, walletDID did.DID, presenta
 	}
 
 	// todo: support multiple wallets
-	vp, err := h.BuildPresentation(ctx, signInstructions[0].VerifiableCredentials, PresentationOptions{
+	vp, err := p.buildPresentation(ctx, &holderDID, signInstructions[0].VerifiableCredentials, PresentationOptions{
 		Format: format,
 		ProofOptions: proof.ProofOptions{
 			Created:   time.Now(),
@@ -137,14 +98,14 @@ func (h wallet) BuildSubmission(ctx context.Context, walletDID did.DID, presenta
 			Expires:   &params.Expires,
 			Nonce:     &params.Nonce,
 		},
-	}, &walletDID, false)
+	})
 	if err != nil {
 		return nil, nil, fmt.Errorf("failed to create verifiable presentation: %w", err)
 	}
 	return vp, &presentationSubmission, nil
 }
 
-func (h wallet) BuildPresentation(ctx context.Context, credentials []vc.VerifiableCredential, options PresentationOptions, signerDID *did.DID, validateVC bool) (*vc.VerifiablePresentation, error) {
+func (p presenter) buildPresentation(ctx context.Context, signerDID *did.DID, credentials []vc.VerifiableCredential, options PresentationOptions) (*vc.VerifiablePresentation, error) {
 	var err error
 	if signerDID == nil {
 		signerDID, err = credential.ResolveSubjectDID(credentials...)
@@ -153,38 +114,29 @@ func (h wallet) BuildPresentation(ctx context.Context, credentials []vc.Verifiab
 		}
 	}
 
-	kid, _, err := h.keyResolver.ResolveKey(*signerDID, nil, resolver.NutsSigningKeyType)
+	kid, _, err := p.keyResolver.ResolveKey(*signerDID, nil, resolver.NutsSigningKeyType)
 	if err != nil {
 		return nil, fmt.Errorf("unable to resolve assertion key for signing VP (did=%s): %w", *signerDID, err)
 	}
-	key, err := h.keyStore.Resolve(ctx, kid.String())
+	key, err := p.keyStore.Resolve(ctx, kid.String())
 	if err != nil {
 		return nil, fmt.Errorf("unable to resolve assertion key from key store for signing VP (did=%s): %w", *signerDID, err)
 	}
 
-	if validateVC {
-		for _, cred := range credentials {
-			err := h.verifier.VerifySignature(cred, &options.ProofOptions.Created)
-			if err != nil {
-				return nil, core.InvalidInputError("invalid credential (id=%s): %w", cred.ID, err)
-			}
-		}
-	}
-
 	switch options.Format {
 	case JWTPresentationFormat:
-		return h.buildJWTPresentation(ctx, *signerDID, credentials, options, key)
+		return p.buildJWTPresentation(ctx, *signerDID, credentials, options, key)
 	case "":
 		fallthrough
 	case JSONLDPresentationFormat:
-		return h.buildJSONLDPresentation(ctx, *signerDID, credentials, options, key)
+		return p.buildJSONLDPresentation(ctx, *signerDID, credentials, options, key)
 	default:
 		return nil, fmt.Errorf("unsupported presentation proof format: %s", options.Format)
 	}
 }
 
 // buildJWTPresentation builds a JWT presentation according to https://www.w3.org/TR/vc-data-model/#json-web-token
-func (h wallet) buildJWTPresentation(ctx context.Context, subjectDID did.DID, credentials []vc.VerifiableCredential, options PresentationOptions, key crypto.Key) (*vc.VerifiablePresentation, error) {
+func (p presenter) buildJWTPresentation(ctx context.Context, subjectDID did.DID, credentials []vc.VerifiableCredential, options PresentationOptions, key crypto.Key) (*vc.VerifiablePresentation, error) {
 	headers := map[string]interface{}{
 		jws.TypeKey: "JWT",
 	}
@@ -216,14 +168,14 @@ func (h wallet) buildJWTPresentation(ctx context.Context, subjectDID did.DID, cr
 	for claimName, value := range options.ProofOptions.AdditionalProperties {
 		claims[claimName] = value
 	}
-	token, err := h.keyStore.SignJWT(ctx, claims, headers, key)
+	token, err := p.keyStore.SignJWT(ctx, claims, headers, key)
 	if err != nil {
 		return nil, fmt.Errorf("unable to sign JWT presentation: %w", err)
 	}
 	return vc.ParseVerifiablePresentation(token)
 }
 
-func (h wallet) buildJSONLDPresentation(ctx context.Context, subjectDID did.DID, credentials []vc.VerifiableCredential, options PresentationOptions, key crypto.Key) (*vc.VerifiablePresentation, error) {
+func (p presenter) buildJSONLDPresentation(ctx context.Context, subjectDID did.DID, credentials []vc.VerifiableCredential, options PresentationOptions, key crypto.Key) (*vc.VerifiablePresentation, error) {
 	ldContext := []ssi.URI{VerifiableCredentialLDContextV1, signature.JSONWebSignature2020Context}
 	ldContext = append(ldContext, options.AdditionalContexts...)
 	types := []ssi.URI{VerifiablePresentationLDType}
@@ -252,119 +204,10 @@ func (h wallet) buildJSONLDPresentation(ctx context.Context, subjectDID did.DID,
 
 	ldProof := proof.NewLDProof(options.ProofOptions)
 	signingResult, err := ldProof.
-		Sign(ctx, document, signature.JSONWebSignature2020{ContextLoader: h.jsonldManager.DocumentLoader(), Signer: h.keyStore}, key)
+		Sign(ctx, document, signature.JSONWebSignature2020{ContextLoader: p.documentLoader, Signer: p.keyStore}, key)
 	if err != nil {
 		return nil, fmt.Errorf("unable to sign VP with LD proof: %w", err)
 	}
 	resultJSON, _ := json.Marshal(signingResult)
 	return vc.ParseVerifiablePresentation(string(resultJSON))
-}
-
-func (h wallet) Put(_ context.Context, credentials ...vc.VerifiableCredential) error {
-	return h.walletStore.put(credentials...)
-}
-
-func (h wallet) List(_ context.Context, holderDID did.DID) ([]vc.VerifiableCredential, error) {
-	return h.walletStore.list(holderDID)
-}
-
-func (h wallet) Remove(ctx context.Context, holderDID did.DID, credentialID ssi.URI) error {
-	err := h.walletStore.remove(holderDID, credentialID)
-	if err == nil {
-		audit.Log(ctx, log.Logger(), audit.VerifiableCredentialRemovedEvent).
-			WithField(core.LogFieldCredentialID, credentialID).
-			WithField(core.LogFieldWalletDID, holderDID).
-			Info("Removed credential from wallet")
-	}
-	return err
-}
-
-func (h wallet) Diagnostics() []core.DiagnosticResult {
-	count, err := h.walletStore.count()
-	if err != nil {
-		log.Logger().WithError(err).Warn("unable to read credential count in wallet")
-	}
-	return []core.DiagnosticResult{
-		core.GenericDiagnosticResult{
-			Title:   "credential_count",
-			Outcome: int(count),
-		},
-	}
-}
-
-func (h wallet) IsEmpty() (bool, error) {
-	count, err := h.walletStore.count()
-	return count == 0, err
-}
-
-var _ schema.Tabler = (*walletRecord)(nil)
-
-type walletRecord struct {
-	HolderDID    string                 `gorm:"primaryKey;column:holder_did"`
-	CredentialID string                 `gorm:"primaryKey"`
-	Credential   store.CredentialRecord `gorm:"foreignKey:CredentialID;references:ID"`
-}
-
-func (walletRecord) TableName() string {
-	return "wallet_credential"
-}
-
-type walletStore struct {
-	db *gorm.DB
-}
-
-func (s walletStore) count() (int64, error) {
-	var count int64
-	err := s.db.Model(walletRecord{}).Count(&count).Error
-	return count, err
-}
-
-func (s walletStore) list(holderDID did.DID) ([]vc.VerifiableCredential, error) {
-	var records []walletRecord
-	err := s.db.Model(walletRecord{}).Preload("Credential").Where("holder_did = ?", holderDID.String()).Find(&records).Error
-	if err != nil {
-		return nil, err
-	}
-	results := make([]vc.VerifiableCredential, 0)
-	for _, record := range records {
-		verifiableCredential, err := vc.ParseVerifiableCredential(record.Credential.Raw)
-		if err != nil {
-			return nil, fmt.Errorf("unable to unmarshal credential %s: %w", record.CredentialID, err)
-		}
-		results = append(results, *verifiableCredential)
-	}
-	return results, nil
-}
-
-func (s walletStore) put(credentials ...vc.VerifiableCredential) error {
-	return s.db.Transaction(func(tx *gorm.DB) error {
-		for _, curr := range credentials {
-			subjectDID, err := curr.SubjectDID()
-			if err != nil {
-				return fmt.Errorf("unable to resolve subject DID from VC %s: %w", curr.ID, err)
-			}
-			record, err := store.CredentialStore{}.Store(tx, curr)
-			if err != nil {
-				return err
-			}
-			if err := tx.FirstOrCreate(&walletRecord{
-				HolderDID:    subjectDID.String(),
-				CredentialID: record.ID,
-			}).Error; err != nil {
-				return err
-			}
-		}
-		return nil
-	})
-}
-
-func (s walletStore) remove(holderDID did.DID, credentialID ssi.URI) error {
-	result := s.db.Where("holder_did = ? AND credential_id = ?", holderDID.String(), credentialID.String()).Delete(&walletRecord{})
-	if result.Error != nil {
-		return result.Error
-	}
-	if result.RowsAffected == 0 {
-		return types.ErrNotFound
-	}
-	return nil
 }
